Correcting Header
Removed  <bits/stdc++.h>

That is not a Standard C++ header file, and Visual C++ does not
implement it. You should not use it even if the compiler you are using
does implement it, as it makes your code immediately non-portable,
possibly even between different versions of the same compiler.